requirements.txt
pybit
pandas
numpy
matplotlib
ta
ccxt
requests
python-dotenv
websocket-client
aiohttp
asyncio
Flask